requirements.txt
pyyaml
pandas
yfinance
pytest
